Uses latest instead of v1alpha3 for JibMavenArtifact and JibGradleArtifact.
