+++ pkg/skaffold/schema/v1alpha3/config.go
@@ -216,10 +216,8 @@ type Profile struct {
 }
 
 type ArtifactType struct {
-	DockerArtifact    *DockerArtifact    `yaml:"docker,omitempty" yamltags:"oneOf=artifact"`
-	BazelArtifact     *BazelArtifact     `yaml:"bazel,omitempty" yamltags:"oneOf=artifact"`
-	JibMavenArtifact  *JibMavenArtifact  `yaml:"jibMaven,omitempty" yamltags:"oneOf=artifact"`
-	JibGradleArtifact *JibGradleArtifact `yaml:"jibGradle,omitempty" yamltags:"oneOf=artifact"`
+	DockerArtifact *DockerArtifact `yaml:"docker,omitempty" yamltags:"oneOf=artifact"`
+	BazelArtifact  *BazelArtifact  `yaml:"bazel,omitempty" yamltags:"oneOf=artifact"`
 }
 
 type DockerArtifact struct {
@@ -231,17 +229,6 @@ type DockerArtifact struct {
 
 type BazelArtifact struct {
 	BuildTarget string `yaml:"target"`
-}
-
-type JibMavenArtifact struct {
-	// Only multi-module
-	Module  string `yaml:"module"`
-	Profile string `yaml:"profile"`
-}
-
-type JibGradleArtifact struct {
-	// Only multi-module
-	Project string `yaml:"project"`
 }
 
 // Parse reads a SkaffoldConfig from yaml.
